Adding validation before delete a category

+++ src/main/java/com/dev/suacomanda/service/CategoryService.java
@@ -1,8 +1,9 @@
 package com.dev.suacomanda.service;
 
 import com.dev.suacomanda.domain.category.Category;
-import com.dev.suacomanda.domain.category.exception.CategoryNotFoundException;
 import com.dev.suacomanda.domain.category.CategoryDTO;
+import com.dev.suacomanda.domain.category.exception.CategoryChildViolationException;
+import com.dev.suacomanda.domain.category.exception.CategoryNotFoundException;
 import com.dev.suacomanda.repositories.CategoryRepository;
 import org.springframework.stereotype.Service;
 
@@ -14,8 +15,12 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    private final ProductService productService;
+
+
+    public CategoryService(CategoryRepository categoryRepository, ProductService productService) {
         this.categoryRepository = categoryRepository;
+        this.productService = productService;
     }
 
     public Category insert(CategoryDTO categoryDTO) {
@@ -32,7 +37,11 @@ public class CategoryService {
     }
 
     public void deleteById(String id) {
-        categoryRepository.deleteById(id);
+        if(productService.findProductsByCategory(id).isEmpty()){
+            categoryRepository.deleteById(id);
+        } else {
+            throw new CategoryChildViolationException("Essa categoria possui produtos atrelados e por isso não pode ser removida");
+        }
     }
 
     public List<Category> findAll() {
